+++ unisannino/denenderman/EntityAIMoveLogsAndPlantablePoint.java
@@ -2,18 +2,13 @@ package unisannino.denenderman;
 
 import java.util.Random;
 
-import net.minecraft.src.Block;
-import net.minecraft.src.BlockSapling;
-import net.minecraft.src.EntityAIBase;
-import net.minecraft.src.Item;
-import net.minecraft.src.ItemSeedFood;
-import net.minecraft.src.ItemSeeds;
-import net.minecraft.src.ItemStack;
-import net.minecraft.src.MathHelper;
-import net.minecraft.src.StepSound;
-import net.minecraft.src.Vec3;
-import net.minecraft.src.World;
-import net.minecraftforge.common.ForgeDirection;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockSapling;
+import net.minecraft.entity.ai.EntityAIBase;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.MathHelper;
+import net.minecraft.util.Vec3;
+import net.minecraft.world.World;
 
 public class EntityAIMoveLogsAndPlantablePoint extends EntityAIBase
 {
@@ -102,9 +97,9 @@ public class EntityAIMoveLogsAndPlantablePoint extends EntityAIBase
 
         for (int var2 = 0; var2 < 10; ++var2)
         {
-            int var3 = MathHelper.floor_double(this.theFarmers.posX + (double)var1.nextInt(20) - 10.0D);
-            int var4 = MathHelper.floor_double(this.theFarmers.boundingBox.minY + (double)var1.nextInt(6) - 3.0D);
-            int var5 = MathHelper.floor_double(this.theFarmers.posZ + (double)var1.nextInt(20) - 10.0D);
+            int var3 = MathHelper.floor_double(this.theFarmers.posX + var1.nextInt(20) - 10.0D);
+            int var4 = MathHelper.floor_double(this.theFarmers.boundingBox.minY + var1.nextInt(6) - 3.0D);
+            int var5 = MathHelper.floor_double(this.theFarmers.posZ + var1.nextInt(20) - 10.0D);
 
 
             if (this.theFarmers.checkIsLogs(this.theWorld, var3, var4, var5) && this.theFarmers.underWoodBlock(this.theWorld, var3, var4, var5) && this.theFarmers.checkTopOfLogs(this.theWorld, var3, var4, var5))
@@ -121,9 +116,9 @@ public class EntityAIMoveLogsAndPlantablePoint extends EntityAIBase
 
         for (int var2 = 0; var2 < 10; ++var2)
         {
-            int var3 = MathHelper.floor_double(this.theFarmers.posX + (double)var1.nextInt(20) - 10.0D);
-            int var4 = MathHelper.floor_double(this.theFarmers.boundingBox.minY + (double)var1.nextInt(6) - 3.0D);
-            int var5 = MathHelper.floor_double(this.theFarmers.posZ + (double)var1.nextInt(20) - 10.0D);
+            int var3 = MathHelper.floor_double(this.theFarmers.posX + var1.nextInt(20) - 10.0D);
+            int var4 = MathHelper.floor_double(this.theFarmers.boundingBox.minY + var1.nextInt(6) - 3.0D);
+            int var5 = MathHelper.floor_double(this.theFarmers.posZ + var1.nextInt(20) - 10.0D);
 
 			ItemStack itemstack = null;
 			Object sapling = null;
